lora-train/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lora-train"
version = "0.1.0"
description = "Local + RunPod LoRA/QLoRA fine-tuning toolkit."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "accelerate>=0.33.0",
  "bitsandbytes>=0.43.0",
  "datasets>=2.20.0",
  "huggingface-hub>=0.24.0",
  "peft>=0.12.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.1",
  "tensorboard>=2.17.0",
  "torch>=2.2.0",
  "transformers>=4.44.0",
  "trl>=0.9.6",
  "wandb>=0.17.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

